extracted everything related to errors to errors.ts
