+++ index.js
@@ -1,8 +1,20 @@
 "use strict";
+var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
+    if (k2 === undefined) k2 = k;
+    Object.defineProperty(o, k2, { enumerable: true, get: function() { return m[k]; } });
+}) : (function(o, m, k, k2) {
+    if (k2 === undefined) k2 = k;
+    o[k2] = m[k];
+}));
+var __exportStar = (this && this.__exportStar) || function(m, exports) {
+    for (var p in m) if (p !== "default" && !exports.hasOwnProperty(p)) __createBinding(exports, m, p);
+};
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.isUTCDateStringFuture = exports.PurchasesErrorHelper = exports.PurchasesErrorCode = exports.INTRO_ELIGIBILITY_STATUS = exports.PACKAGE_TYPE = exports.PRORATION_MODE = exports.PURCHASE_TYPE = exports.ATTRIBUTION_NETWORK = void 0;
+exports.isUTCDateStringFuture = exports.INTRO_ELIGIBILITY_STATUS = exports.PACKAGE_TYPE = exports.PRORATION_MODE = exports.PURCHASE_TYPE = exports.ATTRIBUTION_NETWORK = void 0;
 // @ts-ignore
+var errors_1 = require("./errors");
 var react_native_1 = require("react-native");
+__exportStar(require("./errors"), exports);
 var RNPurchases = react_native_1.NativeModules.RNPurchases;
 var eventEmitter = new react_native_1.NativeEventEmitter(RNPurchases);
 var ATTRIBUTION_NETWORK;
@@ -104,69 +116,6 @@ var INTRO_ELIGIBILITY_STATUS;
      */
     INTRO_ELIGIBILITY_STATUS[INTRO_ELIGIBILITY_STATUS["INTRO_ELIGIBILITY_STATUS_ELIGIBLE"] = 2] = "INTRO_ELIGIBILITY_STATUS_ELIGIBLE";
 })(INTRO_ELIGIBILITY_STATUS = exports.INTRO_ELIGIBILITY_STATUS || (exports.INTRO_ELIGIBILITY_STATUS = {}));
-// Error codes indicating the reson for an error.
-var PurchasesErrorCode;
-(function (PurchasesErrorCode) {
-    PurchasesErrorCode[PurchasesErrorCode["UnknownError"] = 0] = "UnknownError";
-    PurchasesErrorCode[PurchasesErrorCode["PurchaseCancelledError"] = 1] = "PurchaseCancelledError";
-    PurchasesErrorCode[PurchasesErrorCode["StoreProblemError"] = 2] = "StoreProblemError";
-    PurchasesErrorCode[PurchasesErrorCode["PurchaseNotAllowedError"] = 3] = "PurchaseNotAllowedError";
-    PurchasesErrorCode[PurchasesErrorCode["PurchaseInvalidError"] = 4] = "PurchaseInvalidError";
-    PurchasesErrorCode[PurchasesErrorCode["ProductNotAvailableForPurchaseError"] = 5] = "ProductNotAvailableForPurchaseError";
-    PurchasesErrorCode[PurchasesErrorCode["ProductAlreadyPurchasedError"] = 6] = "ProductAlreadyPurchasedError";
-    PurchasesErrorCode[PurchasesErrorCode["ReceiptAlreadyInUseError"] = 7] = "ReceiptAlreadyInUseError";
-    PurchasesErrorCode[PurchasesErrorCode["InvalidReceiptError"] = 8] = "InvalidReceiptError";
-    PurchasesErrorCode[PurchasesErrorCode["MissingReceiptFileError"] = 9] = "MissingReceiptFileError";
-    PurchasesErrorCode[PurchasesErrorCode["NetworkError"] = 10] = "NetworkError";
-    PurchasesErrorCode[PurchasesErrorCode["InvalidCredentialsError"] = 11] = "InvalidCredentialsError";
-    PurchasesErrorCode[PurchasesErrorCode["UnexpectedBackendResponseError"] = 12] = "UnexpectedBackendResponseError";
-    PurchasesErrorCode[PurchasesErrorCode["ReceiptInUseByOtherSubscriberError"] = 13] = "ReceiptInUseByOtherSubscriberError";
-    PurchasesErrorCode[PurchasesErrorCode["InvalidAppUserIdError"] = 14] = "InvalidAppUserIdError";
-    PurchasesErrorCode[PurchasesErrorCode["OperationAlreadyInProgressError"] = 15] = "OperationAlreadyInProgressError";
-    PurchasesErrorCode[PurchasesErrorCode["UnknownBackendError"] = 16] = "UnknownBackendError";
-    PurchasesErrorCode[PurchasesErrorCode["InvalidAppleSubscriptionKeyError"] = 17] = "InvalidAppleSubscriptionKeyError";
-    PurchasesErrorCode[PurchasesErrorCode["IneligibleError"] = 18] = "IneligibleError";
-    PurchasesErrorCode[PurchasesErrorCode["InsufficientPermissionsError"] = 19] = "InsufficientPermissionsError";
-    PurchasesErrorCode[PurchasesErrorCode["PaymentPendingError"] = 20] = "PaymentPendingError";
-    PurchasesErrorCode[PurchasesErrorCode["InvalidSubscriberAttributesError"] = 21] = "InvalidSubscriberAttributesError";
-    PurchasesErrorCode[PurchasesErrorCode["LogOutAnonymousUserError"] = 22] = "LogOutAnonymousUserError";
-})(PurchasesErrorCode = exports.PurchasesErrorCode || (exports.PurchasesErrorCode = {}));
-var PurchasesErrorHelper = /** @class */ (function () {
-    function PurchasesErrorHelper() {
-    }
-    /// Use this to convert an Error to a PurchasesErrorCode.
-    /// It will return `PurchasesErrorCode.UnknownError` if the error code is not
-    /// in the range of PurchasesErrorCodes.
-    ///
-    /// For example: 
-    /// try {
-    ///    PurchaserInfo purchaserInfo = await Purchases.purchasePackage(package);
-    /// } catch (error) {
-    ///     const errorCode = PurchasesErrorHelper.getErrorCode(error);
-    ///     switch(errorCode) {
-    ///     case PurchasesErrorCode.PurchaseCancelledError:
-    ///       print("User cancelled");
-    ///       break;
-    ///     case PurchasesErrorCode.PurchaseNotAllowedError:
-    ///       print("User not allowed to purchase");
-    ///       break;
-    ///     default:
-    ///       // Do other stuff
-    ///       break;
-    ///   }
-    /// }
-    PurchasesErrorHelper.getErrorCode = function (error) {
-        var errorCode = parseInt(error.code);
-        if (errorCode in PurchasesErrorCode) {
-            return errorCode;
-        }
-        else {
-            return PurchasesErrorCode.UnknownError;
-        }
-    };
-    return PurchasesErrorHelper;
-}());
-exports.PurchasesErrorHelper = PurchasesErrorHelper;
 var purchaserInfoUpdateListeners = [];
 var shouldPurchasePromoProductListeners = [];
 exports.isUTCDateStringFuture = function (dateString) {
@@ -320,8 +269,8 @@ var Purchases = /** @class */ (function () {
     Purchases.purchaseProduct = function (productIdentifier, upgradeInfo, type) {
         if (type === void 0) { type = PURCHASE_TYPE.SUBS; }
         return RNPurchases.purchaseProduct(productIdentifier, upgradeInfo, type, null).catch(function (error) {
-            var purchasesErrorCode = PurchasesErrorHelper.getErrorCode(error);
-            error.userCancelled = purchasesErrorCode === PurchasesErrorCode.PurchaseCancelledError;
+            var purchasesErrorCode = errors_1.PurchasesErrorHelper.getErrorCode(error);
+            error.userCancelled = purchasesErrorCode === errors_1.PurchasesErrorCode.PurchaseCancelledError;
             throw error;
         });
     };
@@ -339,8 +288,8 @@ var Purchases = /** @class */ (function () {
             throw new Error("A discount is required");
         }
         return RNPurchases.purchaseProduct(product.identifier, null, null, discount.timestamp.toString()).catch(function (error) {
-            var purchasesErrorCode = PurchasesErrorHelper.getErrorCode(error);
-            error.userCancelled = purchasesErrorCode === PurchasesErrorCode.PurchaseCancelledError;
+            var purchasesErrorCode = errors_1.PurchasesErrorHelper.getErrorCode(error);
+            error.userCancelled = purchasesErrorCode === errors_1.PurchasesErrorCode.PurchaseCancelledError;
             throw error;
         });
     };
@@ -356,8 +305,8 @@ var Purchases = /** @class */ (function () {
      */
     Purchases.purchasePackage = function (aPackage, upgradeInfo) {
         return RNPurchases.purchasePackage(aPackage.identifier, aPackage.offeringIdentifier, upgradeInfo, null).catch(function (error) {
-            var purchasesErrorCode = PurchasesErrorHelper.getErrorCode(error);
-            error.userCancelled = purchasesErrorCode === PurchasesErrorCode.PurchaseCancelledError;
+            var purchasesErrorCode = errors_1.PurchasesErrorHelper.getErrorCode(error);
+            error.userCancelled = purchasesErrorCode === errors_1.PurchasesErrorCode.PurchaseCancelledError;
             throw error;
         });
     };
@@ -375,8 +324,8 @@ var Purchases = /** @class */ (function () {
             throw new Error("A discount is required");
         }
         return RNPurchases.purchasePackage(aPackage.identifier, aPackage.offeringIdentifier, null, discount.timestamp.toString()).catch(function (error) {
-            var purchasesErrorCode = PurchasesErrorHelper.getErrorCode(error);
-            error.userCancelled = purchasesErrorCode === PurchasesErrorCode.PurchaseCancelledError;
+            var purchasesErrorCode = errors_1.PurchasesErrorHelper.getErrorCode(error);
+            error.userCancelled = purchasesErrorCode === errors_1.PurchasesErrorCode.PurchaseCancelledError;
             throw error;
         });
     };
